+++ gestor/settings.py
@@ -66,6 +66,7 @@ INSTALLED_APPS = [
     "schedule",
     "tolls.apps.TollsConfig",
     "menu",
+    "dashboard",
     # Must be the last one
     "rbac",
 ]
