+++ templates/dashboard/business_insights.html
@@ -1,0 +1,56 @@
+{% load static %}
+{% load humanize %}
+<div class="col-lg-12 mb-4 order-0">
+    <div class="card">
+        <div class="d-flex align-items-end row">
+            <div class="col-sm-7">
+                <div class="card-body">
+                    <h5 class="card-title text-primary">
+                        Business Insights <strong>{{ last_date|date:"M, d" }}</strong>
+                    </h5>
+                    <a href="{% url 'weekly-report-date' last_date|date:"mdY" %}"
+                       class="btn btn-sm btn-outline-primary">Week details</a>
+                </div>
+            </div>
+        </div>
+        <div class="card-body row">
+            {% for indicator in indicators %}
+                <div class="col-lg-4 col-md-4 col-6 p-0">
+                    <div class="card m-2">
+                        <div class="card-body">
+                            <div class="card-title d-flex align-items-start justify-content-between">
+                                <div class="avatar flex-shrink-0">
+                                    <img src="{% static indicator.icon %}" alt="chart success" class="rounded">
+                                </div>
+                                <div class="dropdown">
+                                    <button class="btn p-0"
+                                            type="button"
+                                            id="cardOpt3"
+                                            data-bs-toggle="dropdown"
+                                            aria-haspopup="true"
+                                            aria-expanded="false">
+                                        <i class="bx bx-dots-vertical-rounded"></i>
+                                    </button>
+                                    <div class="dropdown-menu dropdown-menu-end"
+                                         aria-labelledby="cardOpt3"
+                                         style="">
+                                        <a class="dropdown-item" href="javascript:void(0);">View More</a>
+                                    </div>
+                                </div>
+                            </div>
+                            <span class="fw-semibold d-block mb-1">{{ indicator.name }}</span>
+                            <h4 class="card-title mb-2">${{ indicator.amount|floatformat:"0"|intcomma }}</h4>
+                            <small class="{% if indicator.positive %}text-success fw-semibold{% else %}text-danger fw-semibold{% endif %}">
+                                {% if indicator.increment > 0 %}
+                                    <i class="bx bx-up-arrow-alt"></i>
+                                {% else %}
+                                    <i class="bx bx-down-arrow-alt"></i>
+                                {% endif %}
+                            {{ indicator.increment|floatformat:"-2" }}%</small>
+                        </div>
+                    </div>
+                </div>
+            {% endfor %}
+        </div>
+    </div>
+</div>
