+++ templates/dashboard/monthly_history.html
@@ -1,0 +1,43 @@
+{% load static %}
+<div class="col-md-6 col-lg-6 mb-4">
+    <div class="card h-100">
+        <div class="card-header">
+            <h5>Monthly history</h5>
+            <ul class="nav nav-pills" role="tablist">
+                {% for indicator in indicators %}
+                    {% if indicator.series %}
+                        <li class="nav-item">
+                            <button type="button"
+                                    class="nav-link {% if forloop.counter0 < 1 %}active{% endif %}"
+                                    role="tab"
+                                    data-bs-toggle="tab"
+                                    data-bs-target="#navs-tabs-line-card-{{ indicator.name }}"
+                                    aria-controls="navs-tabs-line-card-{{ indicator.name }}"
+                                    aria-selected="true">{{ indicator.name }}</button>
+                        </li>
+                    {% endif %}
+                {% endfor %}
+            </ul>
+        </div>
+        <div class="card-body px-0">
+            <div class="tab-content p-0">
+                {% for indicator in indicators %}
+                    {% if indicator.series %}
+                        <div class="tab-pane fade {% if forloop.counter0 < 1 %}show active{% endif %}"
+                             id="navs-tabs-line-card-{{ indicator.name }}"
+                             role="tabpanel"
+                             style="position: relative">
+                            <div id="{{ indicator.name }}Chart"></div>
+                            <div class="resize-triggers">
+                                <div class="expand-trigger">
+                                    <div style="width: 650px; height: 650px;"></div>
+                                </div>
+                                <div class="contract-trigger"></div>
+                            </div>
+                        </div>
+                    {% endif %}
+                {% endfor %}
+            </div>
+        </div>
+    </div>
+</div>
