+++ templates/dashboard/rental_debts.html
@@ -1,0 +1,137 @@
+{% load static %}
+        <div class="col-md-6 col-lg-6 mb-4">
+            <div class="card">
+                <div class="list-head">
+                    <h5 class="card-header">
+                        Rental debt: <strong>${{ rental_debt|floatformat:"0" }}</strong>
+                        <a href="{% url 'client-list' %}"
+                           class="btn btn-sm btn-outline-primary float-end">Details</a>
+                    </h5>
+                    <ul class="nav nav-pills ps-2" role="tablist">
+                        <li class="nav-item">
+                            <button type="button"
+                                    class="nav-link active"
+                                    role="tab"
+                                    data-bs-toggle="tab"
+                                    data-bs-target="#navs-tabs-line-card-date"
+                                    aria-controls="navs-tabs-line-card-date"
+                                    aria-selected="true">By date</button>
+                        </li>
+                        <li class="nav-item">
+                            <button type="button"
+                                    class="nav-link"
+                                    role="tab"
+                                    data-bs-toggle="tab"
+                                    data-bs-target="#navs-tabs-line-card-amount"
+                                    aria-controls="navs-tabs-line-card-amount"
+                                    aria-selected="true">By amount</button>
+                        </li>
+                        <li class="nav-item">
+                            <button type="button"
+                                    class="nav-link"
+                                    role="tab"
+                                    data-bs-toggle="tab"
+                                    data-bs-target="#navs-tabs-line-card-yesterday"
+                                    aria-controls="navs-tabs-line-card-yesterday"
+                                    aria-selected="true">Yesterday</button>
+                        </li>
+                    </ul>
+                </div>
+                <div class="tab-content p-0">
+                    <div class="tab-pane fade show active"
+                         id="navs-tabs-line-card-date"
+                         role="tabpanel">
+                        <div class="table-responsive text-nowrap">
+                            <table class="table">
+                                <thead>
+                                    <tr>
+                                        <th>Name</th>
+                                        <th>Debt</th>
+                                        <th>Overdue</th>
+                                    </tr>
+                                </thead>
+                                <tbody class="table-border-bottom-0">
+                                    {% for associated in clients_by_date %}
+                                        <tr data-search="{{ associated.name }}{{ associated.phone_number }}">
+                                            <td>{% include "users/associated_line.html" %}</td>
+                                            <td>
+                                                <a href="{% url 'client-detail' associated.id %}">${{ associated.debt|floatformat:"2" }}</a>
+                                            </td>
+                                            <td>
+                                                <spam
+                                                {% if associated.overdue %}class="bg-label-danger p-1 rounded"{% endif %}
+                                                >
+                                                {{ associated.last_payment|timesince }}
+                                                </spam>
+                                            </td>
+                                        </tr>
+                                    {% endfor %}
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                    <div class="tab-pane fade" id="navs-tabs-line-card-amount" role="tabpanel">
+                        <div class="table-responsive text-nowrap">
+                            <table class="table">
+                                <thead>
+                                    <tr>
+                                        <th>Name</th>
+                                        <th>Debt</th>
+                                        <th>Overdue</th>
+                                    </tr>
+                                </thead>
+                                <tbody class="table-border-bottom-0">
+                                    {% for associated in clients_by_amount %}
+                                        <tr data-search="{{ associated.name }}{{ associated.phone_number }}">
+                                            <td>{% include "users/associated_line.html" %}</td>
+                                            <td>
+                                                <a href="{% url 'client-detail' associated.id %}">${{ associated.debt|floatformat:"2" }}</a>
+                                            </td>
+                                            <td>
+                                                <spam
+                                                {% if associated.overdue %}class="bg-label-danger p-1 rounded"{% endif %}
+                                                >
+                                                {{ associated.last_payment|timesince }}
+                                                </spam>
+                                            </td>
+                                        </tr>
+                                    {% endfor %}
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                    <div class="tab-pane fade"
+                         id="navs-tabs-line-card-yesterday"
+                         role="tabpanel">
+                        <div class="table-responsive text-nowrap">
+                            <table class="table">
+                                <thead>
+                                    <tr>
+                                        <th>Name</th>
+                                        <th>Debt</th>
+                                        <th>Overdue</th>
+                                    </tr>
+                                </thead>
+                                <tbody class="table-border-bottom-0">
+                                    {% for associated in yesterday_dues %}
+                                        <tr data-search="{{ associated.name }}{{ associated.phone_number }}">
+                                            <td>{% include "users/associated_line.html" %}</td>
+                                            <td>
+                                                <a href="{% url 'client-detail' associated.id %}">${{ associated.debt|floatformat:"2" }}</a>
+                                            </td>
+                                            <td>
+                                                <spam
+                                                {% if associated.overdue %}class="bg-label-danger p-1 rounded"{% endif %}
+                                                >
+                                                {{ associated.last_payment|timesince }}
+                                                </spam>
+                                            </td>
+                                        </tr>
+                                    {% endfor %}
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
