+++ templates/dashboard/trailers_availables.html
@@ -1,0 +1,25 @@
+{% load static %}
+<div class="col-md-6 col-lg-6 mb-4">
+    <div class="card">
+        <div class="list-head">
+            <h5 class="card-header">
+                Trailers available: <strong>{{ available|length }}</strong>
+                <a href="{% url 'list-trailer' %}"
+                   class="btn btn-sm btn-outline-primary float-end">All</a>
+            </h5>
+        </div>
+        <div class="card-body pt-0 scroll trailers-available">
+            {% for equipment in available %}
+                <hr />
+                <div class="d-flex">
+                    {% include "rent/equipment_line.html" with equipment_type="trailer" pinned_image=equipment.pinned_image %}
+                    <a class="btn btn-icon btn-outline-primary"
+                       type="button"
+                       href="{% url 'select-lessee' equipment.id %}">
+                        <span class="tf-icons bx bx-key" style="font-size:1.5em"></span>
+                    </a>
+                </div>
+            {% endfor %}
+        </div>
+    </div>
+</div>
